% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glucocorticoid_exposure.R
\name{glucocorticoid_exposure}
\alias{glucocorticoid_exposure}
\title{Glucocorticoid exposure score calculation for mDNA data}
\usage{
glucocorticoid_exposure(dat)
}
\arguments{
\item{dat}{: preprocessed mDNA data in the GenomicRatioSet-class format (minfi) or with the cpg site information in the rows, samples in the columns.}
}
\value{
glucocorticoid_exposure_score
}
\description{
Provençal, Nadine, Janine Arloth, Annamaria Cattaneo, Christoph Anacker, Nadia Cattane, Tobias Wiechmann, Simone Röh et al.
"Glucocorticoid exposure during hippocampal neurogenesis primes future stress response by inducing changes in DNA methylation."
Proceedings of the National Academy of Sciences 117, no. 38 (2020): 23280-23285.
}
\examples{
data(testdata)
glucocorticoid_exposure(testdata)
}
